rag/requirements.txt
torch == 2.2.2
torchvision>=0.17.0
tqdm==4.66.4
numpy==1.26.4
Pillow==10.3.0
typing-extensions==4.12.2
colpali-engine>=0.3.1,<0.4.0
pymilvus>2.5.0
pymilvus[model]>=2.4.0
sentence-transformers==2.7.0
flask==3.0.3
flask-cors==4.0.1
pdf2image
python-docx
PyMuPDF
langchain
pytest